Removed soft deletion fo the repo since it breaks the HABTM associations
refs #224

Soft deletion of an object using mongoid-paranoia removes the object's id from the HABTM associations which shouldn't happen. These associations have to be created explicitly when restoring the deleted object.
Also, accessing parent association as unscoped fails and is a known bug.

Run 'bundle exec rake utils:restore_repositories RAILS_ENV=production --trace' to restore the deleted repositories.

+++ app/jobs/user_repos_job.rb
@@ -24,15 +24,22 @@ class UserReposJob < ActiveJob::Base
     repo = Repository.unscoped.where(gh_id: gh_repo.id).first
 
     if repo
+      # Commenting the fuctionality since rails 4.2 has an issue with accessing soft deleted parent association. Refer rails issue#10643.
+=begin
       if repo.info.stargazers_count < REPOSITORY_CONFIG['popular']['stars']
-        # soft delete the repository if the star rating has declined.
         repo.set(stars: gh_repo.stargazers_count)
-        repo.destroy
+        # soft delete the repo if it isnt a fork and the star rating has declined.
+        repo.destroy if repo.source_gh_id.nil?
       else
         # restore the repo if the repository was already soft deleted and the current star count is greater then the threshold
-        repo.restore if repo.destroyed?
+        if repo.destroyed?
+          repo.restore
+          # recreate the relationship between repository and its associated users.
+          repo.users.each{|u| u.repositories << repo }
+        end
         repo.set(stars: repo.info.stargazers_count)
       end
+=end
       repo.users << user unless repo.users.include?(user)
       return
     end
